--- AddressBookIO/src/main/java/com/bridzelab/AddressBookIO/Address.java
@@ -1,42 +1,0 @@
-package com.bridzelab.AddressBookIO;
-
-/*
-create variables to set values using getter setter method
- */
-public class Address {
-
-    private String city, state;
-    private long zip;
-
-    public String getCity() {
-        return city;
-    }
-
-    public String getState() {
-        return state;
-    }
-
-    public long getZip() {
-        return zip;
-    }
-
-    public void setCity(String city) {
-        this.city = city;
-    }
-
-    public void setState(String state) {
-        this.state = state;
-    }
-
-    public void setZip(long zip) {
-        this.zip = zip;
-    }
-    /*
-        to accept input and print variable toString method
-     */
-    @Override
-    public String toString() {
-
-        return ", City - "+city+", State - "+state+", Zip Code - "+zip;
-    }
-}
