+++ AddressBookIO/src/main/java/com/bridzelab/AddressBookIO/AddressBookMain.java
@@ -7,7 +7,7 @@ public class AddressBookMain{
 
     public static void main(String[] args) {
         System.out.println("WELCOME TO ADDREES BOOK PROGRAM");
-        HashMap<String, AddressBook> map = new HashMap<>();//Creating HashMap
+        Map<String, AddressBook> map = new HashMap<>();
         System.out.println("Enter Number of address book you want to add");
         int count = scan.nextInt();
         int i = 0;
