--- AddressBookIO/src/main/java/com/bridzelab/AddressBookIO/PersonalDetails.java
@@ -1,107 +1,0 @@
-package com.bridzelab.AddressBookIO;
-
-public class PersonalDetails
-{
-    private String firstName;
-    private String lastName;
-    private String address;
-    private String city;
-    private String state;
-    private Integer zip;
-    private String email;
-    private Long phoneNumber;
-
-
-    public PersonalDetails(String firstName, String lastName, String address, String city, String state, Integer zip, String email, Long phoneNumber) {
-        this.firstName = firstName;
-        this.lastName = lastName;
-        this.address = address;
-        this.city = city;
-        this.state = state;
-        this.zip = zip;
-        this.email = email;
-        this.phoneNumber = phoneNumber;
-    }
-
-    public PersonalDetails()
-    {
-    }
-
-    public void setFirstName(String firstName) {
-        this.firstName = firstName;
-    }
-
-    public void setLastName(String lastName) {
-        this.lastName = lastName;
-    }
-
-    public void setAddress(String address) {
-        this.address = address;
-    }
-
-    public void setCity(String city) {
-        this.city = city;
-    }
-
-    public void setState(String state) {
-        this.state = state;
-    }
-
-    public void setZip(Integer zip) {
-        this.zip = zip;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
-    public void setPhoneNumber(Long phoneNumber) {
-        this.phoneNumber = phoneNumber;
-    }
-
-    public String getFirstName() {
-        return firstName;
-    }
-
-    public String getLastName() {
-        return lastName;
-    }
-
-    public String getAddress() {
-        return address;
-    }
-
-    public String getCity() {
-        return city;
-    }
-
-    public String getState() {
-        return state;
-    }
-
-    public Integer getZip() {
-        return zip;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public Long getPhoneNumber() {
-        return phoneNumber;
-    }
-
-    @Override
-    public String toString()
-    {
-        return "Contacts{" +
-                "firstName='" + firstName + '\'' +
-                ", lastName='" + lastName + '\'' +
-                ", address='" + address + '\'' +
-                ", city='" + city + '\'' +
-                ", state='" + state + '\'' +
-                ", zip=" + zip +
-                ", email='" + email + '\'' +
-                ", phoneNumber=" + phoneNumber +
-                '}';
-    }}
